Classifier Performance on Oversampled Data
Effective techniques using oversampling methods to build classifier for an unbalanced
data set is described in this paper. A dataset is imbalanced if
there are uneven representation of instances from a certain class within the data set.
Typically instances are oversampled or undersampled to improve the sensitivity
of a classifier while predicting the majority class. Synthetic Minority Oversampling
Technique tries to blend under sampling of majority class and oversampling
synthetically generated minority samples.This paper explore the application of
SMOTE and other algorithms which are modified version of SMOTE, such are
SMOTE borderline and ADASYN. By experimenting on a imbalanced dataset, the
performance of a combination of oversampling techniques and classifiers are studied
in this paper.